WIP 3: switch back to using the bash runtimes library
The python library needs some setup of the python path to find the module, while
the bash library just uses some stupid heuristics which can be debugged more
easily.

Right now it can’t find the library because:
1) the target executable gets suffixed with `-repl`, but the runfiles get named
after the name of the rule, which doesn’t have the suffix.
  - We manually remove the suffix (temporary)
2) It looks like the bash library isn’t added to the runfiles?